PALEOocean Reactions

Ocean

Ocean geometry and transport

PALEOocean.add_loop! — Function

add_loop!(A::AbstractMatrix, vm::AbstractVector, L::Real, loopindices)

Add circulation with flux L to transport matrix A, around a closed loop loopindices.

Transport matrix A (s^{-1}) represents tracer transport,

dc/dt = A * c

where c is Vector of tracer concentrations

loopindices is a list of cell indices representing a closed loop, eg [2, 3, 4, 2]

Units for tracers and fluxes are:

Ocean / volume based:

c: mol m-3 tracer concentration

vm: m^3. volume per cell

L: m^3 s-1 (volume flux, cf 1 Sverdrup = 1e6 m^3 s-1)

Atmosphere / mass based

c: kg / total kg, tracer mass mixing ratio

vm: kg, total mass per call

L: kg s-1 (mass flux)

PALEOocean.OceanNoTransport.ReactionOceanNoTransport — Type

ReactionOceanNoTransport

N ocean boxes (default N=1 ie a single 0-D box), each with a surface and floor, no transport.

Parameters

area[Vector{Float64}]=[1.0] (m^2), default_value=[1.0], description="surface / floor area (per box)"

depth[Vector{Float64}]=[1.0] (m), default_value=[1.0], description="depth (per box)"

Methods and Variables

METHODS PALEOboxes.DocStrings.Methods(:methods_setup) exception: UndefVarError(:PALEOreactions)