PALEOocean Reactions

Ocean geometry and transport

PALEOocean.Ocean.OceanNoTransport.ReactionOceanNoTransport — Type

ReactionOceanNoTransport

N ocean boxes (default N=1 ie a single 0-D box), each with a surface and floor, no transport.

Parameters

area[Vector{Float64}]=[1.0] (m^2), default_value=[1.0], description="surface / floor area (per box)"

depth[Vector{Float64}]=[1.0] (m), default_value=[1.0], description="depth (per box)"

Methods and Variables

do_setup_grid

volume (m^3), VT_ReactProperty, description="volume of ocean cells"

volume_total (m^3), VT_ReactProperty, description="total volume of ocean cells"

Abox (m^2), VT_ReactProperty, description="horizontal area of box"

zupper (m), VT_ReactProperty, description="depth of upper surface of box (m) 0 is surface, -100 is depth of 100 m"

zlower (m), VT_ReactProperty, description="depth of lower surface of box (m)"

zmid (m), VT_ReactProperty, description="mean depth of box"

rho_ref (kg m^-3), VT_ReactProperty, description="Ocean transport model density conversion factor (NB: this an artificial model quantity determined by the offline model for tracer transport, and may be distinct from physical ocean density, and from any density anomaly / potl density used by the offline physical model)"

pressure (dbar), VT_ReactProperty, description="Ocean pressure"

Afloor –> oceanfloor.Afloor (m^2), VT_ReactProperty, description="horizontal area of seafloor at base of box"

Afloor_total –> oceanfloor.Afloor_total (m^2), VT_ReactProperty, description="total area of seafloor"

zfloor –> oceanfloor.zfloor (m), VT_ReactProperty, description="depth of ocean floor (m, -ve)"

Asurf –> oceansurface.Asurf (m^2), VT_ReactProperty, description="horizontal area of oceansurface"