PALEOocean Reactions

Ocean geometry and transport

PALEOocean.Ocean.OceanNoTransport.ReactionOceanNoTransport — Type

ReactionOceanNoTransport

N ocean boxes (default N=1 ie a single 0-D box), each with a surface and floor, no transport.

Parameters

area[Vector{Float64}]=[1.0] (m^2), default_value=[1.0], description="surface / floor area (per box)"

depth[Vector{Float64}]=[1.0] (m), default_value=[1.0], description="depth (per box)"

Methods and Variables

do_setup_grid

volume (m^3), VT_ReactProperty, description="volume of ocean cells"

volume_total (m^3), VT_ReactProperty, description="total volume of ocean cells"

Abox (m^2), VT_ReactProperty, description="horizontal area of box"

zupper (m), VT_ReactProperty, description="depth of upper surface of box (m) 0 is surface, -100 is depth of 100 m"

zlower (m), VT_ReactProperty, description="depth of lower surface of box (m)"

zmid (m), VT_ReactProperty, description="mean depth of box"

rho_ref (kg m^-3), VT_ReactProperty, description="Ocean transport model density conversion factor (NB: this an artificial model quantity determined by the offline model for tracer transport, and may be distinct from physical ocean density, and from any density anomaly / potl density used by the offline physical model)"

pressure (dbar), VT_ReactProperty, description="Ocean pressure"

Afloor –> oceanfloor.Afloor (m^2), VT_ReactProperty, description="horizontal area of seafloor at base of box"

Afloor_total –> oceanfloor.Afloor_total (m^2), VT_ReactProperty, description="total area of seafloor"

zfloor –> oceanfloor.zfloor (m), VT_ReactProperty, description="depth of ocean floor (m, -ve)"

Asurf –> oceansurface.Asurf (m^2), VT_ReactProperty, description="horizontal area of oceansurface"

Ocean surface

Air-sea flux

PALEOocean.Oceansurface.AirSeaExchange.ReactionAirSea — Type

ReactionAirSea(Sc_function, sol_function, frac_function, gas_name)

Not called directly - use as ReactionAirSeaO2, ReactionAirSeaCO2, ReactionAirSeaCH4, ReactionAirSeaFixedSolubility

Atmosphere to ocean gas X flux, mol/yr.

PALEOocean.Oceansurface.AirSeaExchange.ReactionAirSeaO2 — Type

ReactionAirSeaO2

See ReactionAirSea

Parameters

solubility_fixed[Bool]=false, default_value=false, description="use fixed solubility"

sol_fix_henry_coeff[Float64]=NaN (mol l-1 atm-1), default_value=NaN, description="Henry's law coefficient for solubility_fixed=true"

moistair[Bool]=true, default_value=true, description="apply correction for moist air"

piston_fixed[Bool]=true, default_value=true, description="use fixed piston velocity"

piston[Float64]=NaN (m d-1), default_value=NaN, description="fixed piston velocity"

TempKmin[Float64]=-Inf (K), default_value=-Inf, description="GENIE bug compatibility - lower limit on temperature for solubility"

Methods and Variables for default Parameters

do_air_sea_flux

Asurf (m^2), VT_ReactDependency, description="horizontal area of oceansurface"

open_area_fraction (), VT_ReactDependency, description="fracton of surface open to atmosphere (0-1.0)"

pXatm –> atm.pO2atm (atm), VT_ReactDependency, description="gas X atmospheric partial pressure"

X_conc –> ocean.oceansurface.O2_conc (mol m-3), VT_ReactDependency, description="ocean concentration [gas X]"

flux_X –> fluxAtmtoOceansurface.flux_O2 (mol yr-1), VT_ReactContributor, description="air -> sea gas X flux"

temp –> ocean.oceansurface.temp (K), VT_ReactDependency, description="ocean surface temperature"

sal –> ocean.oceansurface.sal (psu), VT_ReactDependency, description="ocean salinity"

PALEOocean.Oceansurface.AirSeaExchange.ReactionAirSeaCO2 — Type

ReactionAirSeaO2

See ReactionAirSea

Parameters

solubility_fixed[Bool]=false, default_value=false, description="use fixed solubility"

sol_fix_henry_coeff[Float64]=NaN (mol l-1 atm-1), default_value=NaN, description="Henry's law coefficient for solubility_fixed=true"

moistair[Bool]=true, default_value=true, description="apply correction for moist air"

piston_fixed[Bool]=true, default_value=true, description="use fixed piston velocity"

piston[Float64]=NaN (m d-1), default_value=NaN, description="fixed piston velocity"

TempKmin[Float64]=-Inf (K), default_value=-Inf, description="GENIE bug compatibility - lower limit on temperature for solubility"

PALEOocean.Oceansurface.AirSeaExchange.ReactionAirSeaCH4 — Type

ReactionAirSeaCH4

See ReactionAirSea

Parameters

solubility_fixed[Bool]=false, default_value=false, description="use fixed solubility"

sol_fix_henry_coeff[Float64]=NaN (mol l-1 atm-1), default_value=NaN, description="Henry's law coefficient for solubility_fixed=true"

moistair[Bool]=true, default_value=true, description="apply correction for moist air"

piston_fixed[Bool]=true, default_value=true, description="use fixed piston velocity"

piston[Float64]=NaN (m d-1), default_value=NaN, description="fixed piston velocity"

TempKmin[Float64]=-Inf (K), default_value=-Inf, description="GENIE bug compatibility - lower limit on temperature for solubility"

PALEOocean.Oceansurface.AirSeaExchange.ReactionAirSeaFixedSolubility — Type

ReactionAirSeaFixedSolubility

See ReactionAirSea

Parameters

solubility_fixed[Bool]=true, default_value=true, description="use fixed solubility"

sol_fix_henry_coeff[Float64]=NaN (mol l-1 atm-1), default_value=NaN, description="Henry's law coefficient for solubility_fixed=true"

moistair[Bool]=true, default_value=true, description="apply correction for moist air"

piston_fixed[Bool]=true, default_value=true, description="use fixed piston velocity"

piston[Float64]=NaN (m d-1), default_value=NaN, description="fixed piston velocity"

TempKmin[Float64]=-Inf (K), default_value=-Inf, description="GENIE bug compatibility - lower limit on temperature for solubility"

Methods and Variables for default Parameters

do_air_sea_flux

Asurf (m^2), VT_ReactDependency, description="horizontal area of oceansurface"

open_area_fraction (), VT_ReactDependency, description="fracton of surface open to atmosphere (0-1.0)"

pXatm –> atm.pXatm (atm), VT_ReactDependency, description="gas X atmospheric partial pressure"

X_conc –> ocean.oceansurface.X_conc (mol m-3), VT_ReactDependency, description="ocean concentration [gas X]"

flux_X –> fluxAtmtoOceansurface.flux_X (mol yr-1), VT_ReactContributor, description="air -> sea gas X flux"